Correct float to double comparison
(float) min of 0.05 is 0.0500000007
(double) enteredValue of 0.05 is 0.0500000000
Previous code would block a valid entry. This change may have some unintended consequences.

+++ Firmware/RTK_Everywhere/support.ino
@@ -916,7 +916,7 @@ InputResponse getNewSetting(const char *settingPrompt, float min, float max, flo
 
         if (response == INPUT_RESPONSE_VALID)
         {
-            if (enteredValue >= min && enteredValue <= max)
+            if ((float)enteredValue >= min && enteredValue <= max)
             {
                 *setting = (float)enteredValue; // Recorded to NVM and file at main menu exit
                 return (INPUT_RESPONSE_VALID);
